rag_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
faiss-cpu
sentence-transformers
pypdf
python-dotenv
requests
